+++ src/main/kotlin/codegen/FlutterXPluginInfo.kt
@@ -3,10 +3,10 @@ package codegen
 object FlutterXPluginInfo {
     const val Version: String = "5.3.1.243"
     const val Changelog: String = """
-<h2>5.3.1 - 2025-03-01</h2>
+<h2>5.3.1 - 2025-03-03</h2>
 
-<ul><li>Fix: dart function comment generator tool error (dart comment)</li></ul>
+<ul><li>Fix: dart function comment generator tool error (dart comment)</li><li>Fix: dart assets hover document preview bug (with image) (dart assets preview)</li><li>New: Supports the operation of clicking on the asset path to quickly open files（Need to hold ctrl）</li></ul>
 """
 
-    const val BuildDate: String = "2025-03-01"
+    const val BuildDate: String = "2025-03-03"
 }